+++ src/server/devices/baseDeviceInterfaces/iAcDevice.ts
@@ -2,17 +2,12 @@ import { AcSettings } from '../../../models';
 import { AcMode } from '../../services';
 import { iBaseDevice } from './iBaseDevice';
 
+// TODO: Migrate to new Command-Based System
 export interface iAcDevice extends iBaseDevice {
   settings: AcSettings;
   readonly on: boolean;
   readonly temperature: number;
   readonly mode: AcMode;
-
-  /**
-   * Disable automatic Turn-On for given amount of ms and turn off immediately.
-   * @param {number} timeout
-   */
-  deactivateAutomaticChange(timeout: number): void;
 
   onTemperaturChange(newTemperatur: number): void;